relocate function callbacks code
